Entity listing based on Andreas Solberg metalisting experimental module (http://code.google.com/p/simplesamlphp-labs/source/browse/#svn/trunk/modules/metalisting).
This page shows a list of entities in Janus and validate that these entities have the correct metadata. If the x509 module is installed (http://code.google.com/p/yaco-ssp-modules/source/browse/#svn/trunk/x509) it also checks that the certificates are valid. If the metalisting module is installed it will draw a flag icon next to each entity acording to its tags.

This page is useful for federations since it provides a quick overview of the existing entities and their state. We think this page should be public and no permission should be required to see it.